blah
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@272 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/nrrd/simple.c
@@ -481,7 +481,7 @@ nrrdSanity(void) {
   if (maxsize != sizeof(NRRD_TYPE_BIGGEST)) {
     sprintf(err, "%s: actual max type size is %d != "
 	    "%d == sizeof(NRRD_TYPE_BIGGEST)",
-	    me, maxsize, sizeof(NRRD_TYPE_BIGGEST));
+	    me, maxsize, (int)sizeof(NRRD_TYPE_BIGGEST));
     biffAdd(NRRD, err); return 0;
   }
